+++ client/src/components/mypage/MyRoom/MyRoom.jsx
@@ -1,24 +1,27 @@
 import React from "react";
-import { Thumbnail } from "@components/commons";
+import { useAsync, useModal } from "@hooks";
+import { getMyRoom } from "@api/mypage-api";
+import { Room, EditButton, MyRoomEditModal } from "@components/mypage";
 import styles from "./MyRoom.module.css";
 
-function MyRoom({ roomData }) {
+function MyRoom() {
+  const { data: myRoomData, request: refetchMyRoom } = useAsync({ requestFn: getMyRoom, skip: false });
+  const { isModal, openModal, closeModal } = useModal();
+
   return (
-    <div className={styles.box}>
-      <div className={styles.thumbnail_box}>
-        <Thumbnail />
-      </div>
-      <div className={styles.info_box}>
-        <p className={styles.title}>{roomData.name}</p>
-        <p>
-          {roomData.hashtags.map((hashtag) => (
-            <span>#{hashtag} </span>
-          ))}
-        </p>
-        <p>{roomData.endAt} 까지</p>
-      </div>
-    </div>
+    <>
+      {isModal && <MyRoomEditModal roomData={myRoomData} onClose={closeModal} refetch={refetchMyRoom} />}
+      <section>
+        <div className={styles.heading}>
+          <div>생성 스터디룸</div>
+          {myRoomData && <EditButton onClick={openModal} />}
+        </div>
+        <div className={styles.sub_heading}>생성한 스터디룸</div>
+        <div className={styles.contents}>
+          {myRoomData ? <Room roomData={myRoomData} /> : <div className={styles.no_content}>없습니다.</div>}
+        </div>
+      </section>
+    </>
   );
 }
-
 export default MyRoom;
