+++ client/src/components/mypage/index.js
@@ -2,7 +2,10 @@ export { default as DatePicker } from "./DatePicker/DatePicker";
 export { default as EditButton } from "./EditButton/EditButton";
 export { default as MyGoalEditModal } from "./EditModal/MyGoalEditModal";
 export { default as MyRoomEditModal } from "./EditModal/MyRoomEditModal";
+export { default as MyAccount } from "./MyAccount/MyAccount";
+export { default as MyGoal } from "./MyGoal/MyGoal";
 export { default as MyRoom } from "./MyRoom/MyRoom";
+export { default as Room } from "./Room/Room";
 export { default as SideNavBar } from "./SideNavBar/SideNavBar";
 export { default as StudyTime } from "./StudyTime/StudyTime";
 export { default as TimeTable } from "./TimeTable/TimeTable";
