+++ client/src/pages/MyPage/MyPage.module.css
@@ -15,83 +15,5 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-}
-
-.heading {
-  font-weight: var(--font-semibold);
-  font-size: var(--text-2xl);
-  margin-bottom: 24px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.sub_heading {
-  font-size: var(--text-sm);
-  color: var(--gray600);
-  margin-bottom: 8px;
-}
-
-.contents {
-  margin-bottom: 60px;
-}
-
-.my_goal .contents {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 24px;
-}
-
-.text_box {
-  grid-column: 1/3;
-  font-size: var(--text-sm);
-}
-
-.text_field {
-  height: 90px;
-}
-
-.my_goal .box {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-  margin: 16px;
-  font-size: var(--text-sm);
-}
-
-.bold {
-  font-weight: var(--font-bold);
-  font-size: var(--text-3xl);
-}
-
-.no_content {
-  height: 100px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: var(--gray600);
-  font-size: var(--text-xs);
-}
-
-.account .contents {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.nickname {
-  font-size: var(--text-xl);
-  font-weight: var(--font-medium);
-  margin: 28px 16px;
-}
-
-.delete_btn {
-  background-color: var(--gray900);
-  border: 1px solid rgba(255, 255, 255, 0.1);
-  border-radius: 4px;
-  padding: 12px;
-  color: var(--gray600);
-  margin: 22px 53px;
+  gap: 2rem;
 }
